# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Simple Shop API",
        "description": "A simple shop API",
        "version": "1.0.0"
    },
    "servers": [
        {
          "url": "http://localhost:8080/api/v1",
          "description": "Local server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": ["Product operations"],
                "description": "Returns a list of products",
                "operationId": "findAllProducts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }                                   
                                }                                
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                                
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                              
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": ["Product operations"],
                "description": "Creates a new product",
                "operationId": "createProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "New product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }                                
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                                
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                              
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": ["Product operations"],
                "description": "Returns product by id",
                "operationId": "findProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true 
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }                                
                            }
                        }
                    },
                    "404": {
                        "description": "Product is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                } 
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                                
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                              
                            }
                        }
                    }
                } 
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": ["Product operations"],
                "description": "Updates an existing product",
                "operationId": "updateProduct",
                "parameters": [
                    {                           
                        "name": "id",
                        "in": "path",
                        "required": true                    
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An existing product updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }                                
                            }
                        }   
                    },
                    "404": {
                        "description": "Product is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                } 
                            }
                        }    
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                                
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                              
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": ["Product operations"],
                "description": "Deletes an existing product",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true 
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An existing product deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }                               
                            }
                        }   
                    },
                    "404": {
                        "description": "Product is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                } 
                            }
                        }    
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                                
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                              
                            }
                        }
                    }
                }
            }
        },
        "/products/search": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": ["Product operations"],
                "description": "Returns filtered list of products by name",
                "operationId": "searchProducts",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true 
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }                                   
                                }                                
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                                
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {  
                                    "$ref": "#/components/schemas/Error"                                  
                                }                              
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    }
                },
                "required": ["name", "description", "price"]            
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "statusCode": {
                        "type": "string"
                    }
                }
            }    
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}